# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Diagnostic Center Billing ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List test orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, collected, processing, completed, cancelled]
        - name: patientId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create test order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Bills
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, partial, paid, cancelled]
        - name: patientId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Bills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillList"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create bill for order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get bill by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    put:
      operationId: updateBill
      tags: [bills]
      summary: Update bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillBody"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"

  # Payments
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List referring doctors
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Doctors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorList"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Add referring doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDoctorBody"
      responses:
        "201":
          description: Doctor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"

  # Reports & Analytics
...
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        bloodGroup:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, patientId, firstName, lastName, dateOfBirth, gender, phone, createdAt]
...
    DiagnosticTest:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        duration:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, code, name, category, price, duration, isActive, createdAt]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        patientId:
          type: integer
        patient:
          $ref: "#/components/schemas/Patient"
        doctorId:
          type: integer
          nullable: true
        doctor:
          $ref: "#/components/schemas/Doctor"
          nullable: true
        status:
          type: string
          enum: [pending, collected, processing, completed, cancelled]
        tests:
          type: array
          items:
            $ref: "#/components/schemas/OrderTest"
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        collectedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderNumber, patientId, patient, status, tests, totalAmount, createdAt]
    OrderTest:
      type: object
      properties:
        id:
          type: integer
        testId:
          type: integer
        test:
          $ref: "#/components/schemas/DiagnosticTest"
        price:
          type: number
        result:
          type: string
          nullable: true
        resultStatus:
          type: string
          nullable: true
      required: [id, testId, test, price]
    OrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [orders, total, page, limit]
    CreateOrderBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
          nullable: true
        testIds:
          type: array
          items:
            type: integer
        notes:
          type: string
          nullable: true
      required: [patientId, testIds]
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, collected, processing, completed, cancelled]
        notes:
          type: string
          nullable: true
      required: [status]
    Bill:
      type: object
      properties:
        id:
          type: integer
        billNumber:
          type: string
        orderId:
          type: integer
        order:
          $ref: "#/components/schemas/Order"
        patientId:
          type: integer
        patient:
          $ref: "#/components/schemas/Patient"
        subtotal:
          type: number
        discount:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceAmount:
          type: number
        status:
          type: string
          enum: [draft, pending, partial, paid, cancelled]
        dueDate:
          type: string
          nullable: true
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        createdAt:
          type: string
      required: [id, billNumber, orderId, order, patientId, patient, subtotal, discount, taxAmount, totalAmount, paidAmount, balanceAmount, status, payments, createdAt]
    BillList:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [bills, total, page, limit]
    CreateBillBody:
      type: object
      properties:
        orderId:
          type: integer
        discount:
          type: number
        dueDate:
          type: string
          nullable: true
      required: [orderId]
    UpdateBillBody:
      type: object
      properties:
        discount:
          type: number
        status:
          type: string
          enum: [draft, pending, partial, paid, cancelled]
        dueDate:
          type: string
          nullable: true
    Payment:
      type: object
      properties:
        id:
          type: integer
        billId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, card, upi, insurance, cheque]
        referenceNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, billId, amount, method, createdAt]
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        phone:
          type: string
          nullable: true
        hospitalAffiliation:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, specialization, createdAt]
    DoctorList:
      type: object
      properties:
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/Doctor"
        total:
          type: integer
      required: [doctors, total]
    CreateDoctorBody:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: string
        phone:
          type: string
          nullable: true
        hospitalAffiliation:
          type: string
          nullable: true
      required: [name, specialization]
...