# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Diagnostic Center Billing ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { BillStatus } from "./billStatus";
import type { Order } from "./order";
import type { Patient } from "./patient";
import type { Payment } from "./payment";

export interface Bill {
  id: number;
  billNumber: string;
  orderId: number;
  order: Order;
  patientId: number;
  patient: Patient;
  subtotal: number;
  discount: number;
  taxAmount: number;
  totalAmount: number;
  paidAmount: number;
  balanceAmount: number;
  status: BillStatus;
  dueDate?: string | null;
  payments: Payment[];
  createdAt: string;
}