# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Diagnostic Center Billing ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  patientId: number;
  doctorId?: number | null;
  testIds: number[];
  notes?: string | null;
}