# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Diagnostic Center Billing ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { Doctor } from "./doctor";
import type { OrderStatus } from "./orderStatus";
import type { OrderTest } from "./orderTest";
import type { Patient } from "./patient";

export interface Order {
  id: number;
  orderNumber: string;
  patientId: number;
  patient: Patient;
  doctorId?: number | null;
  doctor?: Doctor | null;
  status: OrderStatus;
  tests: OrderTest[];
  totalAmount: number;
  notes?: string | null;
  collectedAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
}