# lib/api-zod/src/generated/types/orderTest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Diagnostic Center Billing ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosticTest } from "./diagnosticTest";

export interface OrderTest {
  id: number;
  testId: number;
  test: DiagnosticTest;
  price: number;
  result?: string | null;
  resultStatus?: string | null;
}