# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Diagnostic Center Billing ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientGender } from "./patientGender";

export interface Patient {
  id: number;
  patientId: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  gender: PatientGender;
  phone: string;
  email?: string | null;
  address?: string | null;
  bloodGroup?: string | null;
  createdAt: string;
}